pyproject.toml
[tool.poetry]
name = "goit-web-module11"
version = "0.1.0"
description = ""
authors = ["Nataliia Tiutiunnyk <nat.tiutiunnyk@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.109.2"
uvicorn = "^0.27.1"
sqlalchemy = "^2.0.27"
psycopg2-binary = "^2.9.9"
alembic = "^1.13.1"
dotenv = "^0.0.5"
python-dotenv = "^1.0.1"
asyncpg = "^0.29.0"
pydantic-extra-types = "^2.5.0"
email-validator = "^2.1.0.post1"
phonenumbers = "^8.13.30"
libgravatar = "^1.0.4"
python-multipart = "^0.0.9"
python-jose = "^3.3.0"
passlib = "^1.7.4"
fastapi-jwt-auth = "^0.5.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
